rename files to snake_case; update requirements.txt
