+++ requirements.txt
@@ -1,33 +1,8 @@
-async-generator==1.10
-attrs==22.1.0
-certifi==2022.9.24
-charset-normalizer==2.1.1
-exceptiongroup==1.0.0rc9
-execnet==1.9.0
-h11==0.14.0
-idna==3.4
-iniconfig==1.1.1
-outcome==1.2.0
-packaging==21.3
-pluggy==1.0.0
-py==1.11.0
-pyparsing==3.0.9
-PySocks==1.7.1
 pytest==7.1.3
-pytest-forked==1.4.0
 pytest-html==3.1.1
-pytest-metadata==2.0.2
 pytest-xdist==2.5.0
 python-decouple==3.6
 python-dotenv==0.21.0
 requests==2.28.1
 selenium==4.5.0
-sniffio==1.3.0
-sortedcontainers==2.4.0
-tomli==2.0.1
-tqdm==4.64.1
-trio==0.22.0
-trio-websocket==0.9.2
-urllib3==1.26.12
 webdriver-manager==3.8.3
-wsproto==1.2.0
